requirements.txt
fastapi
uvicorn
pymupdf
python-dotenv
google-generativeai
pydantic